+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [1.5.0] - 2026-09-21
+
 ### Added
 
 - `uniservice` is on [PyPI](https://pypi.org/project/uniservice/). The READMEs recommend